+++ unlock-js/CHANGELOG.md
@@ -6,7 +6,7 @@
 - Adding ability to check if address is key granter
 - Adding ability to cancel key by key owner
 - Adding ability to cancel key by lock owner
-- Adding ability to transfer key by key owner
+- Adding ability to share key by key owner
 - Adding ability to retrieve key manager for key
 
 # 0.16.3